ai-services/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1


# Google Cloud Vertex AI SDK for Gemini
google-cloud-aiplatform==1.30.0
# newer version to interact with Gemini
google-generativeai==0.8.5


# OpenAI Python client for GPT-4
openai==1.14.3

# Anthropic client (Claude 3)
anthropic==0.3.1

# Transformers and tokenizers for CodeLlama
transformers==4.33.2
tokenizers==0.13.3

requests==2.31.0


torch==2.2.2
